Adding Anandita's code for her
Manually replacing Anandita's old categories.py with Anandita's updated categories.py, because she does not have access to github.

Please note, it appears that this categories.py code will NOT WORK because there are function calls that have not imported the proper files and do not use the appropriate syntax.  I am simply uploading it for her.